Refactor createPaginationNumeration as a hook (usePagination).
